+++ doc/changelog.rst
@@ -17,6 +17,8 @@ Update
   Compared to handwritten implementation, the implementation from previous version is slower about 1% ~ 3%,
   and even worse when returning a value type with expensive copy construction, such as a string with extremely huge content.
 
+- Experimental support for enum with custom underlying type in AngelScript 2.39 WIP.
+
 - ``script_invoke_result`` now supports ``operator->`` for pointer type,
   which is useful for accessing members of returned reference types.